requirements.txt
symbol-sdk-python==3.1.0
python-dotenv==1.0.0
websockets==11.0.3
requests==2.31.0